+++ lib/src/service_provider.dart
@@ -16,8 +16,8 @@ abstract class ServiceProvider {
   /// If the service does not exist, null is returned.
   T? tryResolve<T>();
 
-  /// Checks if a with the [T] or [t] is registered.
-  bool has<T>([Type? t]);
+  /// Checks if a with the [T] or [type] is registered.
+  bool has<T>([Type? type]);
 
   /// Boot the service container.
   /// While booting the service provider, preloaded services are instantiated.