+++ run_tests.sh
@@ -2,9 +2,9 @@
 set -e
 
 cd ./example
-pub get
+dart pub get
 rm -f lib/src/example.container.dart
-pub run build_runner build --delete-conflicting-outputs
+dart pub run build_runner build --delete-conflicting-outputs
 sed -i'' "s+package\:catalyst_builder_example\/src\/++g" "lib/src/example.container.dart"
 sed -i'' "s+package\:third_party_dependency\/+\.\.\/\.\.\/\.\.\/test\/third_party_dependency\/lib\/+g" "lib/src/example.container.dart"
 cd ..
@@ -13,8 +13,8 @@ dart test
 
 REPO_TOKEN=$1
 if [ "$REPO_TOKEN" ]; then
-  pub global activate dart_coveralls
-  pub global run dart_coveralls report \
+ dart pub global activate dart_coveralls
+ dart pub global run dart_coveralls report \
     --token $REPO_TOKEN \
     --retry 2 \
     --exclude-test-files \