+++ test/unit/annotation/preload_test.dart
@@ -4,6 +4,6 @@ import 'package:test/test.dart';
 void main() {
   test('Preload constructor', () {
     const preload = Preload();
-    expect(preload, TypeMatcher<Preload>());
+    expect(preload, const TypeMatcher<Preload>());
   });
 }