+++ test/unit/annotation/service_test.dart
@@ -1,11 +1,10 @@
 import 'package:catalyst_builder/catalyst_builder.dart';
-import 'package:catalyst_builder/src/annotation/annotation.dart';
 import 'package:test/test.dart';
 
 void main() {
   test('Service Constructor', () {
     const simpleService = Service();
-    expect(simpleService, TypeMatcher<Service>());
+    expect(simpleService, const TypeMatcher<Service>());
 
     expect(simpleService.exposeAs, isNull);
     expect(simpleService.lifetime, ServiceLifetime.singleton);
@@ -18,7 +17,7 @@ void main() {
       lifetime: ServiceLifetime.transient,
       exposeAs: String,
     );
-    expect(serviceWithLifetimeAndExposeAs.exposeAs, TypeMatcher<Type>());
+    expect(serviceWithLifetimeAndExposeAs.exposeAs, const TypeMatcher<Type>());
     expect(serviceWithLifetimeAndExposeAs.exposeAs, equals(String));
     expect(serviceWithLifetimeAndExposeAs.lifetime, ServiceLifetime.transient);
   });